adding files to match names PE-GWAS and PE-GEA
The variable names used in the PEs were different in some cases from
the GWAS and the GEAS.  One issue, for the GWAS, was the presence of
two root-shoot variables and lack of a match for total weight.  A
different issue, for the GEAs, was that some variables (MAP and MSP)
were log transformed for the PEs but not the GEAs.