Working code, pine & spruce, corrected vs raw
This code visually compares genetic architectures for phenotypes in
specific environments for pine and spruce, and compares the values
corrected for pop structure vs. the raw, uncorrected values.  The raw
values show the kinds of interesting patterns that we might expect.
The corrected values make no sense… it might be still a problem in the
sign of the SNPs not being calculated appropriately?
Still need to fix plotting of scores…