Changed outlier criteria to be based on P-values
I have completely revised the methods and about half the results. The
major change has been in the criteria that I was using to obtain
outlier SNPs - using the new criteria is more in line with what Sam did
in the Science paper and should also give more similar results to what
Jeremy is getting with the array data. This changes a lot of the
details of the results but not the big picture.
Other major changes include doing all the stats to justify the
groupings and adding background about the galaxy plot concept.

The next major steps are to:

1) Choose ~10-20 top SNPs for Jon to plot on the landscape - I could
use some feedback on how to choose these. Jon can also plot a composite
allele frequency, but I think that would involve re-coding all the
alleles so they are in low frequency in one type of environment and
high frequency in another type of environment.

2) Confirm with Mike W that the bootstrap approach for significance
testing of the covariance is statistically valid.

3) Redo the annotations and GO enrichment with the new list of outliers.